+++ src/data structures/QueryResult.cs
@@ -12,14 +12,14 @@ public readonly struct QueryResult
     /// <summary>
     /// Gets any user-defined flags to distinguish the gen index. 
     /// </summary>
-    public readonly byte Flag;
+    public readonly int Flag;
 
     /// <summary>
     /// Constructs a QueryResult.
     /// </summary>
     /// <param name="genIndex">the associated gen index.</param>
     /// <param name="flag">any user-defined flags to ditinguish the gen index.</param>
-    public QueryResult(GenIndex genIndex, byte flag)
+    public QueryResult(GenIndex genIndex, int flag)
     {
         GenIndex = genIndex;
         Flag = flag;
